requirements.txt
django-cors-headers
djangorestframework
markdown
django-filter
qrcode[pil]